Markdown documentation with code snippets.
Now when `markdown` is the target language, instead of generating a page model
in the literate programming style of Docco, the page is parsed as GitHub
flavored markdown.

Edify will detect the "```" escape sequence that defines a code snippet. If the
escape sequence is followed by the name of a Pygments language highlighter, the
code snippet will be highlighted using Pygments. If the Edify `Cakefile` for the
site defines a divider, then all snippets of the same language are run through
Pygments at the same time using the divider comment method from Docco.

If there is no language, the source is formatted as plain text. XML special
characters are escaped and the text is wrapped in a `pre` element.

Also...

Fixed `say` and `die`. The first parameter to `apply`, the value of `this` for
the function call, is obviously the `console` itself and not `null`. This works
fine in Node.js, but when I copied and pasted the code into source evaluated by
Firefox, I got errors. I think Firefox. Might have been Chrome. It's wrong in
any case.