Created `Makefile`.
Removed Twinkie and the `Cakefile`. Created a `Makefile` instead.

This was brought over from Stencil, which I've just altered to use GNU
`make` instead of `cake`. Had to create new recipes for compiling
Streamlined CoffeeScript into JavaScript libraries.

The Makefile uses `inotify-tools` to watch the source directories and
rebuild using `make` if anything changes. Builds are faster than they
were with the `Cakefile` because of the dependency resolution of `make`.

You'll need to `yum install inotify-tools` to enjoy `make watch`.

I'm able to create rules that are specific to a set of targets using
[4.12.1 Syntax of Static Pattern
Rules](http://www.gnu.org/software/make/manual/html_node/Static-Usage.html).
This was a problem at first. A basic rule based on a wildcarded target
for the JavaSript like `lib/%.js` can have only one recipe, so it can be
either the vendor JavaScript copy or the Streamlined CoffeeScript
compile. Whichever you choose, the other would need to be spelled out
explicitly; target, requisites and recipe. Static patterns allow you to
specify different rules for different sets of targets.

Note that, I make `lib` directory creation a separate target on which
the vendor JavaScript and Streamlined CoffeeScript depended. However,
this dirtied the contents of the `lib` directory. When the vendor
JavaScript was updated, the modified time of the `lib` directory was
updated, making the Streamlined CoffeeScript out of date. The
Streamlined CoffeeScript is then updated in the next run, updating the
modified time of the `lib` directory, making the vendor JavaScript out
of date. It would resolve itself eventually somehow, but only after two
pointless builds.