Convert to Stencil supporting library.
Converted to a library for use with Stencil. Edify will be a collection
of utilities useful for generating collections to use with Stencil when
generating GitHub:Pages.

Create a first test of the `docco` function against a CoffeeScript file.

Removed all code that would cause Edify to search for files, run files
through Stencil, serve files via HTTP.